Trying to get the stats a bit more consistent.

+++ src/populationNodes/Agents/Agent.java
@@ -7,8 +7,8 @@ import PopulationModel.Node;
 public interface Agent extends Node {
 	
 	//Fitness Calculation
-	public void setFitness(int fitness);
-	public int getFitness();
+	public void setFitness(double fitness);
+	public double getFitness();
 	public void adjustFinalFitnessValue();
 	
 	//Statistics
